デフォルト鍵を表示するボタンを配置
